Create init handler
