+++ src/tracer/src/cli/helper.rs
@@ -1,4 +1,3 @@
-use std::fs::File;
 use std::io;
 use std::io::Write;
 use std::process::Command;
